Adds server mode.
